feat: Hide taint error messages, order error messages by priority
